desktop/pyproject.toml
[project]
name = "pdfQaGemini"
version = "0.1.0"
description = "PDF Q&A with Gemini - Desktop MVP"
requires-python = ">=3.10"
dependencies = [
    "PySide6>=6.6.0",
    "qasync>=0.24.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "google-genai>=0.2.0",
    "pymupdf>=1.23.0",
    "pillow>=10.1.0",
    "cachetools>=5.3.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.3.0",
    "boto3>=1.34.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "black>=23.0.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
